Update MIT License and project file references
Modified the LICENSE file to update the MIT license, removing the year 2025 from the copyright notice and adjusting the formatting of some lines. The disclaimer section was retained with minor formatting changes.

In Sufficit.Net.Http.csproj, changed the license file reference from "LICENSE" to "license," possibly indicating a change in naming convention or how the file is referenced in the project.